Watchdog to reset if HTTP request locks device up.
Starts the Arduino watchdog timer (with an 8 second timeout) before making the HTTP request to the API, so that the device will be reset if it crashes. The watchdog timer is disabled after the connection is closed.

Testing has shown that the device can sometimes freeze after making a request.